Refactor: use modern java for
I personally like having an index variable to see and use for things
but I also don't like seeing inteliji bitching at me about how I can use
an "enhanced for loop" for this.

Yes, inteliji I know I CAN, but do I WANT to? Not really. I think the
syntactic sugar of converting

  for (int i = 0; i < length; i++)

to the incomprehensable

  for (String arg : args)

is less readable from a declarative point of view. When I see the latter
I _know_ that it's processing the data as an iteratable thing and giving
me the item but that doenst mean I don't like the verboseness of seeing
that the thing is an array that you get stuf out via index. If I wanted
to have less verbose code I wouldn't be using Java now would I.