Create Consumer to update path extension
A filename visitor is easy enough to use, but if I want to not have to
pass in the extensions and other arguments to the visitor then I'll need
to provide it will a simple void operation via a Consumer.

This ones name is a little bit tricky since it really does represent an
action to be done, which makes me want to call it RenameAction or maybe
ChangeExtensionAction. But since it IS a consumer subclass I also want
to put that into the name of the class to make it obvious. Tricky.