Add generic Path visitor
Give it a consumer to run and it will go forth and hit every path you
ask it to with the consumer action. If given a recursion boolean of true
then it will go down the path, we could probably make it relatively
simple to make the checks against whether or not it should accept a file
and run the consumer or not, same with the way the recursiveness is
checked.

If we wanted to make it so that it doesn't other with hidden files or
that sort of thing, then we'd probably want to make it behave that way.