Add initial GUI
This has actually existed the whole time, just not commited yet as we
worked through the other gui elements and pieces.

I get the feeling I'll look at this code in 3 months and think "Huh,
that's kinda gross" but for now, it is what it is and we'll keep it as
it is.