Create + use JustTheMessageFormatter
The default formatter for the handlers in java.util.logging will print
off the error level and the class and method it was called form. Which
is really cool, but also, not what I want when it comes to printing out
my errors to the log.

Specifically I don't want this because it makes the --help text say
Optional ifPresent which just feels silly. So let's ditch all the useful
 information about the log record and just show its message instead for
 that one thing. Yay.